+++ data-raw/preprocess_data.R
@@ -1100,6 +1100,15 @@ parameters_SharmaParton2007_hd <- readr::read_csv(
     AIC = as.numeric(AIC)
   )
 
+parameters_Rijal2012_hd <- readr::read_csv(
+  "data-raw/parameters_Rijal2012_hd.csv",
+  show_col_types = FALSE
+) %>%
+  mutate(
+    Species = standardize_species_code(Species, keep_all = FALSE),
+    R2 = as.numeric(R2)
+  )
+
 # combine all into one ####
 
 internal_objs <- c(
@@ -1129,6 +1138,7 @@ internal_objs <- c(
   "parameters_Huang1994_si",
   "parameters_Huang2013_hd",
   "parameters_SharmaParton2007_hd",
+  "parameters_Rijal2012_hd",
   "parameters_Carmean1989",
   "parameters_Carmean1996",
   "parameters_QC_IQS2013",